pyproject.toml
[tool.poetry]
name = "nginx-security-monitor"
version = "1.0.0"
description = "A simple set of Python scripts to monitor NGINX logs, detect attack patterns, mitigate intrusion tactics, and trigger alerts through email and SMS."
authors = ["AccessiTech <accessit3ch@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/AccessiTech/nginx-security-monitor"
repository = "https://github.com/AccessiTech/nginx-security-monitor"
keywords = ["nginx", "security", "monitoring", "logs", "alerts"]

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.25.1"
pyyaml = "^5.4.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"